requirements.txt
about-time==4.2.1
absl-py==1.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aiotg==2.0.0
alive-progress==3.1.4
altair==5.4.0
altgraph==0.17.3
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
auto-py-to-exe==2.32.0
beautifulsoup4==4.12.2
bidict==0.23.1
billiard==4.2.0
blinker==1.8.2
bottle==0.12.25
bottle-websocket==0.2.9
branca==0.7.1
bs4==0.0.1
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.1
cairocffi
CairoSVG==2.7.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
cleo==2.0.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
comtypes==1.4.5
contourpy==1.0.7
crashtest==0.4.1
cron-descriptor==1.4.0
cryptography==41.0.5
cssselect2==0.7.0
cycler==0.11.0
dacite==1.8.1
DateTime==5.2
debugpy==1.8.1
decorator==5.1.1
deepl==1.17.0
defusedxml==0.8.0rc2
diff-match-patch==20230430
distlib==0.3.6
Django==4.2.7
django-allauth==0.58.1
django-celery-beat==2.5.0
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-grappelli==3.0.8
django-import-export==3.3.4
django-js-asset==2.1.0
django-modeltranslation==0.18.11
django-rangefilter==0.9.1
django-rosetta==0.9.9
django-timezone-field==6.1.0
djangorestframework==3.14.0
docopt==0.6.2
docutils==0.19
drf-yasg==1.21.7
dulwich==0.21.3
easygoogletranslate==0.0.4
Eel==0.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==2.0.1
fastapi==0.109.2
filelock==3.12.0
Flask==3.0.3
flatbuffers==23.5.26
folium==0.15.1
fonttools==4.39.2
frozenlist==1.4.0
fsspec==2023.12.2
future==0.18.3
gast==0.4.0
gevent==22.10.2
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.22.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grapheme==0.6.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
h5py==3.1.0
html5lib==1.1
htmlmin==0.1.12
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
huggingface-hub==0.20.2
iconify==0.0.103
idna==3.4
ifaddr==0.2.0
ImageHash==4.3.1
imbalanced-learn==0.12.3
imblearn==0.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
installer==0.7.0
ipykernel==6.29.4
ipython==8.18.1
itsdangerous==2.2.0
jaraco.classes==3.2.3
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
Js2Py==0.74
jsonschema==4.17.3
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.15.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==23.13.1
kids.cache==0.0.7
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kivymd==1.2.0
kiwisolver==1.4.4
kombu==5.3.4
labelImg==1.8.6
libclang==16.0.0
libsass==0.21.0
llvmlite==0.42.0
lockfile==0.12.2
lxml==5.1.0
Markdown==3.4.3
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
minio==7.1.15
ml-dtypes==0.2.0
mock==5.0.1
more-itertools==9.1.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multimethod==1.11.2
Naked==0.1.32
narwhals==1.4.1
nest-asyncio==1.6.0
networkx==3.2.1
nicegui==1.4.22
nltk==3.8.1
numba==0.59.1
numpy==1.24.3
oauthlib==3.2.2
odfpy==1.4.1
opencv-contrib-python==4.10.0.84
opencv-python==4.8.0.76
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.10.1
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
pandas-profiling==3.6.6
parso==0.8.4
patsy==0.5.6
pefile==2023.2.7
pexpect==4.8.0
phik==0.12.4
pika==1.3.2
Pillow==10.0.1
pipwin==0.5.2
pkginfo==1.9.6
platformdirs==2.6.2
playwright==1.41.0
pluggy==1.2.0
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.1
polib==1.2.0
progressbar2==4.2.0
prompt-toolkit==3.0.41
protobuf==4.23.4
pscript==0.7.7
psutil==5.9.7
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyaes==1.6.1
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycaw==20240210
pycparser==2.21
pycryptodome==3.11.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.9.1
pydotplus==2.0.2
pyee==11.0.1
PyGetWindow==0.0.9
Pygments==2.17.2
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.0
pyjokes==0.6.0
pyjsparser==2.7.1
PyJWT==2.8.0
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyparsing==3.0.9
pypdf==4.0.1
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
PyPrind==2.11.3
pyproject_hooks==1.0.0
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyqt5-tools==5.15.4.3.2
PyQtChart==5.15.6
PyQtChart-Qt5==5.15.2
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.30
PySide2==5.15.2.1
pySmartDL==1.3.4
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.4.0
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-engineio==4.9.0
python-multipart==0.0.9
Python-Rest-Framework==0.3.14
python-socketio==5.11.2
python-utils==3.8.1
python3-openid==3.2.0
pytweening==1.2.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.6.0
pywhatkit==5.4
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==6.0.1
pyzmq==26.0.3
qrcode==7.4.2
QT-PyQt-PySide-Custom-Widgets==0.6.4
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
QtPy==2.3.0
qtsass==0.3.2
rapidfuzz==2.15.1
redis==5.0.1
regex==2023.12.25
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.7.1
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.12.2
selenium==4.17.2
sentencepiece==0.2.0
shellescape==3.8.1
shellingham==1.5.0.post1
shiboken2==5.15.2.1
simple-colors==0.1.5
simple-websocket==1.0.0
six==1.15.0
smmap==5.0.1
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sortedcontainers==2.4.0
sounddevice==0.4.7
soupsieve==2.4
spark-parser==1.8.9
SQLAlchemy==2.0.6
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.36.3
statsmodels==0.14.1
streamlit==1.37.1
svglib==1.5.1
sympy==1.12
tablib==3.5.0
Telethon==1.33.1
tenacity==8.5.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-gpu==2.5.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
term-background==1.0.1
termcolor==1.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
TikTokApi==6.2.0
tiktokapipy==0.2.4.post2
tinycss2==1.2.1
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
torch==2.1.2
torchvision==0.16.2
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.3
transformers==4.37.2
trio==0.24.0
trio-websocket==0.11.1
trove-classifiers==2023.4.25
typeguard==4.2.1
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==4.3
ultralytics==8.0.229
uncompyle6==3.9.1
uritemplate==4.1.1
uritools==4.0.2
urllib3==1.26.15
uvicorn==0.22.0
vbuild==0.8.2
vine==5.1.0
virtualenv==20.16.5
visions==0.7.6
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.12
webdriver-manager==4.0.1
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
whichcraft==0.6.1
wikipedia==1.4.0
wordcloud==1.9.3
wrapt==1.12.1
wsproto==1.2.0
xdis==6.1.0
xgboost==2.1.1
xhtml2pdf==0.2.14
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.0
xyzservices==2023.10.1
yarl==1.9.2
ydata-profiling==4.7.0
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
